Let script write to stdout, redirect to output file.
The output of "translate_sample_sheet_for_report" is generated in the
current working directory with respect to the script.  Depending on
how it is executed, the file may end up in an unexpected location,
causing Snakemake to wonder where the file went.  The confusion
disappears when writing to standard output and redirecting explicitly
to the output file.  This ensures that the working directory is
irrelevant.