Canonicalize all file arguments to the report script.
When given relative file names the report script may not be able to
find files, so we use readlink to canonicalize all file names.  In the
common case this is not needed, because we use absolute file names
throughout.  For remote execution we may opt to delay canonicalization
to avoid retaining references to local directories.